requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
Pillow==9.2.0
requests==2.28.1
urllib3==1.26.11
Werkzeug==2.2.2
